@@ -14,7 +14,7 @@
 	<name>danhom1</name>
 	<description>Project MyDrive</description>
 	<properties>
-		<java.version>17</java.version>
+		<java.version>21</java.version>
 	</properties>
 	<dependencies>
 		<dependency>
