@@ -1,4 +1,4 @@
-package com.example.danhom1.User;
+package com.example.danhom1.Model.User;
 
 import com.example.danhom1.Exception.UserAlreadyExistedException;
 import lombok.AllArgsConstructor;
@@ -12,13 +12,13 @@ import org.springframework.transaction.annotation.Transactional;
 public class UserService {
     private final UserRepo userRepo;
 //    private final UserStorageRepo userStorageRepo;
-    public User RegisterNewUser(@NonNull User user) throws UserAlreadyExistedException {
+    public void RegisterNewUser(@NonNull User user) throws UserAlreadyExistedException {
         if (EmailExists((user.getEmail())))
             throw new UserAlreadyExistedException();
 //        UserStorage userStorage = new UserStorage();
 //        userStorage.setUser(user);
 //        userStorageRepo.save(userStorage);
-        return userRepo.save(user);
+        userRepo.save(user);
     }
 
     private Boolean EmailExists(String email){
